Fikser feil i opprettelse av fagsystemId
* Ved "fiksing" av deprecated Char.toByte() kom jeg i skade for å endre til Char.code, som er en Int, men det skulle vært Char.code.toByte()

+++ sykepenger-model/src/main/kotlin/no/nav/helse/utbetalingstidslinje/UtbetalingsreferanseBuilder.kt
@@ -10,7 +10,7 @@ internal fun genererUtbetalingsreferanse(uuid: UUID): String {
 
 private fun UUID.base32Encode(): String {
     val pad = '='
-    return Base32(pad.code)
+    return Base32(pad.code.toByte())
         .encodeAsString(this.byteArray())
         .replace(pad.toString(), "")
 }
